refactor(api): address review — consistent pins, terser docstrings
- Loud effect doctest (camas/__init__) and the effect-plugin fixture now
  import every name from camas.v0 submodules instead of mixing root and
  pinned imports — consistent pinning, since the contract types aren't on
  the root surface anyway.
- camas.v0 module docstring reduced to a single line.
- tests/test_v0.py module docstring removed (tests are D-exempt).
- root package versioning note tightened.
